+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **[NEW]** Implemented `Call::argument()` ([#56]).
 - **[NEW]** Implemented `MockBuilder::source()` for easier debugging ([#45]).
 - **[IMPROVED]** Vast improvements to verification failure output ([#66]).
+- **[IMPROVED]** Allow use of phonySelf parameter everywhere ([#63]).
 - **[IMPROVED]** Optimizations to the matcher driver system ([#67]).
 - **[IMPROVED]** Optimizations to the equal to matcher ([#69]).
 - **[IMPROVED]** Calls to eval() no longer use @ suppression.
@@ -15,6 +16,7 @@
 [#17]: https://github.com/eloquent/phony/issues/17
 [#45]: https://github.com/eloquent/phony/issues/45
 [#56]: https://github.com/eloquent/phony/issues/56
+[#63]: https://github.com/eloquent/phony/issues/63
 [#66]: https://github.com/eloquent/phony/issues/66
 [#67]: https://github.com/eloquent/phony/issues/67
 [#69]: https://github.com/eloquent/phony/issues/69